# packages/contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: commerce-ai-ops-platform API
  version: 0.1.0
  description: >-
    REST v1. Every domain endpoint requires a Bearer ES256 JWT for audience commerce-api.
    Tenant and subject come from the verified token; the role comes from the membership row.
    Locally a development issuer signs tokens (pnpm auth:init); cloud uses an OIDC provider (M11).
  license:
    name: UNLICENSED
    identifier: UNLICENSED
servers:
  - url: http://localhost:3001
    description: Local development
security:
  - bearerAuth: []
tags:
  - name: operations
  - name: catalog
  - name: orders
  - name: inventory
  - name: support
  - name: agents
paths:
  /healthz:
    get:
      operationId: getLiveness
      security: []
      tags: [operations]
      summary: Process liveness
      responses:
        '200':
          description: Process is running
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '../schemas/health-report.schema.json'
  /readyz:
    get:
      operationId: getReadiness
      security: []
      tags: [operations]
      summary: Readiness including PostgreSQL, pgvector availability and Redis
      responses:
        '200':
          description: All dependencies reachable
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '../schemas/health-report.schema.json'
        '503':
          description: At least one dependency is unreachable
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '../schemas/health-report.schema.json'
  /v1/status:
    get:
      operationId: getServiceStatus
      security: []
      tags: [operations]
      summary: Diagnostic status echoing the correlation id
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: API reachable
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '../schemas/service-status.schema.json'
  /v1/products:
    get:
      operationId: listProducts
      summary: List published SKUs with filters and cursor pagination
      tags: [catalog]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: category, in: query, schema: { type: string } }
        - { name: currency, in: query, schema: { type: string, enum: [USD] } }
        - {
            name: price_lt,
            in: query,
            description: 'Strict upper bound in minor units (price_minor < price_lt)',
            schema: { type: integer, minimum: 0 },
          }
        - { name: ram_gb, in: query, schema: { type: integer, minimum: 0 } }
        - { name: cpu_family, in: query, schema: { type: string, maxLength: 128 } }
        - { name: region, in: query, schema: { type: string } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50 } }
      responses:
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '200':
          description: Page of published SKUs
          content:
            application/json:
              schema:
                $ref: '../schemas/product-list.schema.json'
        '400':
          $ref: '#/components/responses/Error'
  /v1/products/{id}:
    get:
      operationId: getProduct
      summary: Get a product and its published SKUs
      tags: [catalog]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '200':
          description: Product and published SKUs
          content:
            application/json:
              schema:
                $ref: '../schemas/product-list.schema.json'
        '404':
          $ref: '#/components/responses/Error'
  /v1/inventory/{sku}:
    get:
      operationId: getInventory
      summary: Get availability for a SKU
      tags: [inventory]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: sku, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: region, in: query, schema: { type: string } }
      responses:
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '200':
          description: Availability
          content:
            application/json:
              schema:
                $ref: '../schemas/inventory.schema.json'
        '404':
          $ref: '#/components/responses/Error'
  /v1/orders/{id}:
    get:
      operationId: getOrder
      summary: Get an order owned by the caller
      tags: [orders]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Order with immutable snapshots
          content:
            application/json:
              schema:
                $ref: '../schemas/order.schema.json'
        '404':
          $ref: '#/components/responses/Error'
  /v1/orders/{id}/shipping:
    get:
      operationId: getShippingStatus
      summary: Get per-package shipping status
      tags: [orders]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Per-package shipping status
          content:
            application/json:
              schema:
                $ref: '../schemas/shipping-status.schema.json'
        '404':
          $ref: '#/components/responses/Error'
  /v1/anomalies:
    get:
      operationId: listAnomalies
      summary: List deduplicated inventory anomalies
      tags: [inventory]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: Deduplicated inventory alerts
  /v1/support-tickets:
    post:
      operationId: createSupportTicket
      summary: Create a support ticket with explicit consent
      tags: [support]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - {
            name: Idempotency-Key,
            in: header,
            required: true,
            schema: { type: string, minLength: 8, maxLength: 128 },
          }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [customer_id, category, summary, confirmed]
              properties:
                customer_id: { type: string, format: uuid }
                order_id: { type: string, format: uuid }
                category: { type: string }
                summary: { type: string }
                confirmed: { type: boolean }
      responses:
        '201':
          description: Ticket created or replayed
        '400':
          $ref: '#/components/responses/Error'
  /v1/action-requests:
    post:
      operationId: createActionRequest
      summary: Request a privileged order action
      tags: [orders]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
      responses:
        '201':
          description: Action request pending approval
  /v1/approvals/{id}/decision:
    post:
      operationId: decideApproval
      summary: Approve or reject an action request
      tags: [orders]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Decision recorded
  /v1/orders/{id}/actions:
    post:
      operationId: executeOrderAction
      summary: Execute an approved order action
      tags: [orders]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Cancellation requested
        '409':
          $ref: '#/components/responses/Error'
  /v1/agent-runs:
    post:
      operationId: createAgentRun
      summary: Start an agent run
      tags: [agents]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '202':
          description: Run accepted
          content:
            application/json:
              schema:
                $ref: '../schemas/agent-run.schema.json'
  /v1/agent-runs/{id}:
    get:
      operationId: getAgentRun
      summary: Get durable agent run state
      tags: [agents]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Durable run state
          content:
            application/json:
              schema:
                $ref: '../schemas/agent-run.schema.json'
  /v1/agent-runs/{id}/events:
    get:
      operationId: streamAgentRunEvents
      summary: Stream agent run events (SSE)
      tags: [agents]
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: Last-Event-ID, in: header, schema: { type: string } }
      responses:
        '200':
          description: SSE of durable run events
components:
  parameters:
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{8,64}$'
  headers:
    CorrelationId:
      description: Correlation id used in structured logs and trace attributes.
      schema:
        type: string
        pattern: '^[A-Za-z0-9._-]{8,64}$'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Stable domain error
      content:
        application/json:
          schema:
            $ref: '../schemas/error.schema.json'